app/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
ollama
dotenv
PyPDF2
python-docx
python-multipart
sqlalchemy
python-jose
passlib
psycopg2-binary